fix(tests): add conftest_optional import for core deps
